+++ composeApp/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [v3.0.0] - TODO
 - remove insets
-- `Move` - can download all media for a char
+- `Move` 
+  - can download all media for a char
+  - basic character info
 - fixed Kimberly
 
 ## [v2.3.0] - 2026-08-23